// src/api/resources/plans/client/requests/PlansSubscribeRequest.ts
// This file was auto-generated by Fern from our API Definition.

/**
 * @example
 *     {
 *         customerExternalId: "customerExternalId",
 *         currency: "currency"
 *     }
 */
export interface PlansSubscribeRequest {
    /** The external ID of the customer */
    customerExternalId: string;
    /** The currency code for the order (e.g., "USD", "EUR") */
    currency: string;
}