// src/api/resources/plans/client/requests/PlansUpgradeRequest.ts
// This file was auto-generated by Fern from our API Definition.

/**
 * @example
 *     {
 *         customerExternalId: "customerExternalId",
 *         oldPlanId: "oldPlanId",
 *         newPlanId: "newPlanId"
 *     }
 */
export interface PlansUpgradeRequest {
    /** The external ID of the customer */
    customerExternalId: string;
    /** The ID of the current plan */
    oldPlanId: string;
    /** The ID of the plan to upgrade to */
    newPlanId: string;
}